fix(task-5): seperate accounts and crud apps
