+++ project5/accounts/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+from . import models
+
+# Register your models here.
+admin.site.register(models.UserProfile)
