+++ project5/accounts/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class AccountsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'accounts'
+    def ready(self):
+        import accounts.signals
