+++ project5/accounts/templates/accounts/base.html
@@ -1,0 +1,113 @@
+{% load static %}
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>{% block title %}{% endblock %}</title>
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/css/bootstrap.min.css" rel="stylesheet" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.1/font/bootstrap-icons.css">
+    <link rel="stylesheet" href="{% static 'css/dark.css' %}">
+    <style>
+        body {
+            background-color: #1e1e2e; /* nord/orchis compatible dark */
+            color: #cdd6f4;
+        }
+        .navbar {
+            background-color: #181825 !important; 
+        }
+        .navbar .nav-link,
+        .navbar-brand {
+            color: #cdd6f4 !important;
+        }
+        .navbar .nav-link:hover {
+            color: #89b4fa !important; /* nord blue hover */
+        }
+
+        .dropdown-menu {
+            background-color: #181825;
+        }
+        .dropdown-menu .dropdown-item {
+            color: #cdd6f4;
+        }
+        .dropdown-menu .dropdown-item:hover {
+            background-color: #45475a;
+        }
+
+        a {
+            color: #89b4fa;
+        }
+        a:hover {
+            color: #b4befe;
+        }
+
+        .btn-primary {
+            background-color: #89b4fa;
+            border-color: #89b4fa;
+            color: #1e1e2e;
+        }
+        .btn-primary:hover {
+            background-color: #74a0e0;
+            border-color: #74a0e0;
+        }
+    </style>
+</head>
+
+<body>
+    <nav class="navbar navbar-expand-lg navbar-dark">
+        <div class="container-fluid">
+            
+            <a href="{% url 'homepage' %}" class="navbar-brand">CRUD</a>
+
+            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#nav-links">
+                <span class="navbar-toggler-icon"></span>
+            </button>
+
+            <div class="collapse navbar-collapse" id="nav-links">
+                
+                <ul class="navbar-nav">
+                    <li class="nav-item"><a href="{% url 'developers_list' %}" class="nav-link">Devs List</a></li>
+                    <li class="nav-item"><a href="{% url 'projects_list' %}" class="nav-link">Projects List</a></li>
+                    <li class="nav-item"><a href="{% url 'new_developer' %}" class="nav-link">New Dev</a></li>
+                    <li class="nav-item"><a href="{% url 'new_project' %}" class="nav-link">New Project</a></li>
+                </ul>
+
+                <ul class="navbar-nav ms-auto">
+                    {% if user.is_authenticated %}
+                        <li class="nav-item">
+                            <a class="nav-link" href="{% url 'dashboard' %}">
+                                <i class="bi bi-person-circle"></i> {{ user.username }}
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-danger" href="{% url 'confirm_logout' %}">
+                                <i class="bi bi-box-arrow-right"></i> Logout
+                            </a>
+                        </li>
+                    {% else %}
+                        <li class="nav-item">
+                            <a class="nav-link" href="{% url 'login' %}">
+                                <i class="bi bi-box-arrow-in-right"></i> Login
+                            </a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link" href="{% url 'signup' %}">
+                                <i class="bi bi-person-plus"></i> Sign Up
+                            </a>
+                        </li>
+                    {% endif %}
+                </ul>
+
+            </div>
+        </div>
+    </nav>
+
+    <div class="container mt-4">
+        {% block content %}{% endblock %}
+    </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
+</body>
+</html>
