+++ project5/accounts/templates/accounts/password_reset_done.html
