+++ project5/crud/admin.py
@@ -5,4 +5,4 @@ from . import models
 admin.site.register(models.Developer)
 admin.site.register(models.Project)
 admin.site.register(models.Skill)
-admin.site.register(models.UserProfile)
+
