+++ project5/crud/apps.py
@@ -4,6 +4,5 @@ from django.apps import AppConfig
 class CrudConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'crud'
-    def ready(self):
-        import crud.signals
+
