+++ project5/crud/urls.py
@@ -25,41 +25,5 @@ urlpatterns = [
 
     #path('error/user-not-found/', views.user_not_found, name='user_not_found'),
 
-    path('user/signup/', views.NewUser.as_view(), name='signup'),
-    path('user/login/', views.LoginUser.as_view(), name='login'),
-    path('user/logout-confirm/', views.ConfirmLogout.as_view(), name='confirm_logout'),
-    path('user/logout/', views.LogoutUser.as_view(), name='logout'),
-    path('user/profile/edit', views.UpdateUserProfile.as_view(), name='update_profile'),
-    path('user/dashboard', views.Dashboard.as_view(), name='dashboard'),
 
-    path('user/reset-password/',
-        auth_views.PasswordResetView.as_view(
-            template_name='crud/password_reset.html',
-            email_template_name='crud/password_reset_email.html',
-            subject_template_name='crud/password_reset_subject.txt',
-            success_url='/user/reset-password/done/'
-        ),
-        name='password_reset'),
-
-    
-    path('user/reset-password/done/',
-        auth_views.PasswordResetDoneView.as_view(
-            template_name='crud/password_reset_done.html'
-        ),
-        name='password_reset_done'),
-
-    
-    path('user/reset/<uidb64>/<token>/',
-        auth_views.PasswordResetConfirmView.as_view(
-            template_name='crud/password_reset_confirm.html',
-            success_url='/user/reset/complete/'
-        ),
-        name='password_reset_confirm'),
-
-    
-    path('user/reset/complete/',
-        auth_views.PasswordResetCompleteView.as_view(
-            template_name='crud/password_reset_complete.html'
-        ),
-        name='password_reset_complete'),
 ]
